Moving compnent to common and adding tests
